s2: Disable use of the AOSP hwcomposer 2-over-1 adapter
Reason for revert: The HWC2 composer uses vndservicemanager for
managing services. This is used by the libqservice which registers
when hwc starts. The service libqservice is used by mm-pp-daemon and
the camera driver. With no modification, these components expect
the service at the system servicemanager instead of
the vndservicemanager. So they aren't able to connect and
prints out

ServiceManager: Waiting for service display.qservice...

at the logfile.
As workaround removing the TARGET_USES_HWC2 flag revert to the
old hwc1 composer, which registers services at
the system servicemanager.

Change-Id: I4f360add2de05a4b9d8d7046bfcb9c9cb5008108